re-organize pymc wrapper into a main module and utils
